+++ R/citation.R
@@ -15,6 +15,9 @@
 #'   console; ignored if it is \code{NULL})
 #' @param tweak whether to fix some known problems in the citations, especially
 #'   non-standard format of authors
+#' @param prefix a prefix string for keys in BibTeX entries; by default, it is
+#'   \samp{R-} unless \code{\link{option}('knitr.bib.prefix')} has been set to
+#'   another string
 #' @return a list containing the citations (also written to the \code{file} as a
 #'   side effect)
 #' @note Some packages on CRAN do not have standard bib entries, which was once
@@ -33,15 +36,19 @@
 #'   packages. See \code{knitr:::.tweak.bib} for details of tweaks. Also note
 #'   this is subject to future changes since R packages are being updated.
 #' @export
+#' @author Yihui Xie and Michael Friendly
 #' @examples write_bib(c('RGtk2', 'gWidgets'), file = 'R-GUI-pkgs.bib')
 #' write_bib(c('animation', 'rgl', 'knitr', 'ggplot2'))
 #' write_bib(c('base', 'parallel', 'MASS'))  # base and parallel are identical
+#' write_bib('cluster', prefix = '')  # a empty prefix
+#' write_bib('digest', prefix = 'R-pkg-')  # a new prefix
 #' write_bib(c('rpart', 'survival'))
 #' write_bib(c('rpart', 'survival'), tweak = FALSE) # original version
 #'
 #' # what tweak=TRUE does
 #' str(knitr:::.tweak.bib)
-write_bib = function(x = .packages(), file = '', tweak = TRUE) {
+write_bib = function(x = .packages(), file = '', tweak = TRUE,
+                     prefix = getOption('knitr.bib.prefix', 'R-')) {
   idx = mapply(system.file, package = x) == ''
   if (any(idx)) {
     warning('package(s) ', paste(x[idx], collapse = ', '), ' not found')
@@ -51,7 +58,7 @@ write_bib = function(x = .packages(), file = '', tweak = TRUE) {
   bib = sapply(x, function(pkg) {
     cite = citation(pkg, auto = if (pkg == 'base') NULL else TRUE)
     entry = toBibtex(cite)
-    entry[1] = sub('\\{,$', sprintf('{R-%s,', pkg), entry[1])
+    entry[1] = sub('\\{,$', sprintf('{%s%s,', prefix, pkg), entry[1])
     gsub('', '', entry)
   }, simplify = FALSE)
   if (tweak) {